feat: move InstallamentChooserItem class to PaymentMethodResources
