+++ sdk/src/main/java/co/omise/android/ui/PaymentMethodResources.kt
@@ -128,3 +128,53 @@ sealed class PaymentChooserItem(
             get() = PaymentChooserItem::class.nestedClasses.mapNotNull { it.objectInstance as? PaymentChooserItem }
     }
 }
+
+internal sealed class InstallmentChooserItem(
+        @DrawableRes override val iconRes: Int,
+        override val title: String? = null,
+        @StringRes override val titleRes: Int? = null,
+        @DrawableRes override val indicatorIconRes: Int
+) : OmiseListItem {
+    companion object
+    object Bbl : InstallmentChooserItem(
+            iconRes = R.drawable.payment_bbl,
+            titleRes = R.string.payment_method_installment_bbl_title,
+            indicatorIconRes = R.drawable.ic_next
+    )
+
+    object KBank : InstallmentChooserItem(
+            iconRes = R.drawable.payment_kasikorn,
+            titleRes = R.string.payment_method_installment_kasikorn_title,
+            indicatorIconRes = R.drawable.ic_next
+    )
+
+    object Bay : InstallmentChooserItem(
+            iconRes = R.drawable.payment_bay,
+            titleRes = R.string.payment_method_installment_bay_title,
+            indicatorIconRes = R.drawable.ic_next
+    )
+
+    object FirstChoice : InstallmentChooserItem(
+            iconRes = R.drawable.payment_first_choice,
+            titleRes = R.string.payment_method_installment_first_choice_title,
+            indicatorIconRes = R.drawable.ic_next
+    )
+
+    object Ktc : InstallmentChooserItem(
+            iconRes = R.drawable.payment_ktc,
+            titleRes = R.string.payment_method_installment_ktc_title,
+            indicatorIconRes = R.drawable.ic_next
+    )
+
+    object Scb : InstallmentChooserItem(
+            iconRes = R.drawable.payment_scb,
+            titleRes = R.string.payment_method_installment_scb_title,
+            indicatorIconRes = R.drawable.ic_next
+    )
+
+    data class Unknown(val bankName: String) : InstallmentChooserItem(
+            iconRes = R.drawable.payment_installment,
+            title = bankName,
+            indicatorIconRes = R.drawable.ic_next
+    )
+}
